backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.7.0
pdfminer.six>=20231228
python-docx>=1.1.0
pandas>=2.2.0
python-multipart>=0.0.7
